Rename Jackson IT tests to Jackson instead of JSon
