+++ vertx-core/src/test/java/io/vertx/it/jackson/JacksonReadConstraintsOverrideTest.java
@@ -9,28 +9,18 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
 
-package io.vertx.it.json;
+package io.vertx.it.jackson;
 
-import io.vertx.core.json.DecodeException;
-import io.vertx.core.json.JsonObject;
-import io.vertx.test.core.VertxTestBase;
+import io.vertx.tests.json.jackson.JacksonReadConstraintsTestBase;
 import org.junit.Test;
 
 /**
  * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
  */
-public class NoDatabindTest extends VertxTestBase {
+public class JacksonReadConstraintsOverrideTest {
 
   @Test
-  public void testJsonObject() {
-    JsonObject obj = new JsonObject("{\"foo\":\"bar\"}");
-    assertEquals("bar", obj.getString("foo"));
-    assertEquals("{\"foo\":\"bar\"}", obj.toString());
-    try {
-      obj.mapTo(Object.class);
-      fail();
-    } catch (DecodeException ignore) {
-      // Expected
-    }
+  public void testReadConstraints() {
+    JacksonReadConstraintsTestBase.testReadConstraints(100,  100, 100, 100, 1000);
   }
 }
